The first attempt
I know some of the code is more bloated where it doesn't need to be. I may come back to this and make it cleaner. The use of positive look ahead was very helpful here. And it can be used in the future to consolidate some of the RegEx compile statements.